+++ mpsl/include/mpsl.h
@@ -26,6 +26,7 @@ extern "C" {
 #include "nrf_peripherals.h"
 #include "nrf_errno.h"
 #include "mpsl_clock.h"
+#include "mpsl_hwres.h"
 
 /** @brief High IRQ priority
  *
@@ -36,40 +37,6 @@ extern "C" {
 
 /** @brief Size of build revision array in bytes. */
 #define MPSL_BUILD_REVISION_SIZE 20
-
-/* Reserved (D)PPI, PPIB and IPCT resources for the supported platforms. */
-#if defined(NRF52_SERIES)
-#define MPSL_RESERVED_PPI_CHANNELS ((1UL << 19) | (1UL << 30) | (1UL << 31))
-/* This corresponds to the PPI channels 19, 30, and 31. */
-#define MPSL_PPI_CHANNELS_USED_MASK (0xc0080000)
-#elif defined(NRF53_SERIES)
-#define MPSL_RESERVED_PPI_CHANNELS ((1UL << 0) | (1UL << 1) | (1UL << 2))
-/* This corresponds to the DPPI channels 0, 1, and 2. */
-#define MPSL_DPPIC_CHANNELS_USED_MASK (0x00000007)
-#elif defined(NRF54L_SERIES) || defined(NRF71_SERIES)
-#define MPSL_RESERVED_PPI_CHANNELS (1UL << 0)
-#define MPSL_DPPIC10_CHANNELS_USED_MASK (0x00000001)
-#define MPSL_DPPIC20_CHANNELS_USED_MASK (0x00000001)
-#define MPSL_PPIB11_CHANNELS_USED_MASK  (0x00000001)
-#define MPSL_PPIB21_CHANNELS_USED_MASK  (0x00000001)
-#elif defined(NRF54H_SERIES) || defined(GRTC_PRESENT)
-#define MPSL_RESERVED_PPI_CHANNELS (1UL << 0)
-#define MPSL_DPPIC020_CHANNELS_USED_MASK (0x00000001)
-#define MPSL_IPCT130_CHANNELS_USED_MASK  (0x00000001)
-#else
-#error Unknown NRF series.
-#endif
-
-/* Defines which timer is being used by the MPSL implementation */
-#if defined(NRF52_SERIES)
-	#define MPSL_TIMER0 NRF_TIMER0
-#elif defined(NRF53_SERIES)
-	#define MPSL_TIMER0 NRF_TIMER0_NS
-#elif defined(NRF54L_SERIES) || defined(NRF71_SERIES)
-	#define MPSL_TIMER0 NRF_TIMER10
-#else
-	#define MPSL_TIMER0 NRF_TIMER020
-#endif
 
 /** @brief    Function prototype for the assert handler.
  *